refactor: split code
