+++ src/components/Widgets/index.ts
@@ -1,5 +1,5 @@
-export * from './Widget';
 export * from './Widget.style';
+export * from './Widget.types';
 export * from './WidgetEvents';
 export * from './Widgets';
 export * from './Widgets.style';
